+++ doc/user/content/release-notes.md
@@ -48,6 +48,12 @@ Wrap your release notes at the 80 character mark.
 
 {{% version-header v0.7.2 %}}
 
+- Add a [`--log-filter` command-line option](/cli/#logging) and a
+  `MZ_LOG_FILTER` environment variable that control which log messages to emit.
+
+  This behavior was previously available via the undocumented `MZ_LOG`
+  environment variable, which will be removed in a future release.
+
 - Record Kafka Consumer metrics in the `mz_kafka_consumer_statistics` system
   table. Enabled by default for all Kafka sources.